+++ src/qibolab/instruments/rfsoc/__init__.py
@@ -1,0 +1,3 @@
+"""RFSoC module driver for qibosoq."""
+
+from .driver import RFSoC
